+++ lib/ArangoDBClient/CollectionHandler.php
@@ -246,9 +246,7 @@ class CollectionHandler extends Handler
      *                          <p>Options are:<br>
      *                          <li>'type'                 - 2 -> normal collection, 3 -> edge-collection</li>
      *                          <li>'waitForSync'          - if set to true, then all removal operations will instantly be synchronised to disk / If this is not specified, then the collection's default sync behavior will be applied.</li>
-     *                          <li>'journalSize'          - journalSize value.</li>
      *                          <li>'isSystem'             - false->user collection(default), true->system collection .</li>
-     *                          <li>'isVolatile'           - false->persistent collection(default), true->volatile (in-memory) collection .</li>
      *                          <li>'keyOptions'           - key options to use.</li>
      *                          <li>'distributeShardsLike' - name of prototype collection for identical sharding.</li>
      *                          <li>'numberOfShards'       - number of shards for the collection.</li>
@@ -257,6 +255,7 @@ class CollectionHandler extends Handler
      *                          <li>'shardKeys'            - array of shard key attributes.</li>
      *                          <li>'shardingStrategy'     - sharding strategy to use in cluster.</li>
      *                          <li>'smartJoinAttribute'   - attribute name for smart joins (if not shard key).</li>
+     *                          <li>'schema'               - collection schema.</li>
      *                          </p>
      *
      * @return mixed - id of collection created
@@ -275,16 +274,8 @@ class CollectionHandler extends Handler
             $collection->setWaitForSync($this->getConnectionOption(ConnectionOptions::OPTION_WAIT_SYNC));
         }
 
-        if ($collection->getJournalSize() === null) {
-            $collection->setJournalSize($this->getConnectionOption(ConnectionOptions::OPTION_JOURNAL_SIZE));
-        }
-
         if ($collection->getIsSystem() === null) {
             $collection->setIsSystem($this->getConnectionOption(ConnectionOptions::OPTION_IS_SYSTEM));
-        }
-
-        if ($collection->getIsVolatile() === null) {
-            $collection->setIsVolatile($this->getConnectionOption(ConnectionOptions::OPTION_IS_VOLATILE));
         }
 
         $type   = $collection->getType() ?: Collection::getDefaultType();
@@ -292,10 +283,9 @@ class CollectionHandler extends Handler
             Collection::ENTRY_NAME         => $collection->getName(),
             Collection::ENTRY_TYPE         => $type,
             Collection::ENTRY_WAIT_SYNC    => $collection->getWaitForSync(),
-            Collection::ENTRY_JOURNAL_SIZE => $collection->getJournalSize(),
             Collection::ENTRY_IS_SYSTEM    => $collection->getIsSystem(),
-            Collection::ENTRY_IS_VOLATILE  => $collection->getIsVolatile(),
             Collection::ENTRY_KEY_OPTIONS  => $collection->getKeyOptions(),
+            Collection::ENTRY_SCHEMA       => $collection->getSchema(),
         ];
 
         // set extra cluster attributes
@@ -326,6 +316,7 @@ class CollectionHandler extends Handler
         if ($collection->getSmartJoinAttribute() !== null) {
             $params[Collection::ENTRY_SMART_JOIN_ATTRIBUTE] = $collection->getSmartJoinAttribute();
         }
+
 
         $response = $this->getConnection()->post(Urls::URL_COLLECTION, $this->json_encode_wrapper($params));