+++ CHANGELOG.md
@@ -22,7 +22,6 @@ In addition, the following functionality is deprecated:
 
 - CollectionHandler::load()
 - CollectionHandler::unload()
-- 
 
 The `Cursor` class will now fetch outstanding cursor result data via HTTP POST requests to
 `/_api/cursor/<cursor-id>`. It previously fetched further results via HTTP PUT requests from
@@ -36,6 +35,14 @@ The driver now supports the following options for document CRUD operations:
 - "overwriteMode"
 - "silent"
 - "keepNull", "mergeObjects" (for insert API if overwriteMode=update)
+
+The following options have been removed in class Collection:
+- isVolatile
+- journalSize
+
+The following functions have been removed in class Collection:
+- setJournalSize(), getJournalSize()
+- setIsVolatile(), getIsVolatile()
 
 ## Release notes for the ArangoDB-PHP driver 3.7.x